WORKSPACE: ensure that project's dependencies are loaded first
go_repository targets are loaded in several places in this project:
- gazelle_dependencies
- goblet_deps

When a lib is declared in both, which everone is declared first would be
the one used by the compiler.

Adjust the loading order so that goblet_deps takes priority.
This will help dealing with some golang.org/x/foobar modules in future
commits.